+++ lib/storage/secure_storage/token_storage.dart
@@ -1,7 +1,10 @@
+import 'dart:async';
+
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
 class TokenStorage {
   final FlutterSecureStorage _storage;
+  final _tokenStreams = <String, StreamController<String?>>{};
 
   new(this._storage);
 
@@ -15,6 +18,7 @@ class TokenStorage {
   ) async {
     await _storage.write(key: _access(userId), value: access);
     await _storage.write(key: _refresh(userId), value: refresh);
+    _tokenStreams[userId]?.add(access);
   }
 
   Future<String?> getAccessToken(String userId) =>
@@ -26,10 +30,23 @@ class TokenStorage {
   Future<void> clearTokens(String userId) async {
     await _storage.delete(key: _access(userId));
     await _storage.delete(key: _refresh(userId));
+    await _tokenStreams[userId]?.close();
+    _tokenStreams.remove(userId);
   }
 
   Future<bool> hasAccessToken(String userId) async {
     final token = await _storage.read(key: _access(userId));
     return token != null && token.isNotEmpty;
+  }
+
+  Stream<String?> tokenStream(String userId) {
+    return (_tokenStreams[userId] ??= StreamController<String?>.broadcast(
+      onCancel: () {
+        if (_tokenStreams[userId]?.hasListener == false) {
+          _tokenStreams[userId]?.close();
+          _tokenStreams.remove(userId);
+        }
+      },
+    )).stream;
   }
 }